% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spamAnalysis.R
\name{buildModel}
\alias{buildModel}
\title{Building model for Spam Analysis.}
\usage{
buildModel(spamData)
}
\arguments{
\item{spamData}{a data frame with the words specified, and one column named type where is stored if spam or not}
}
\value{
a model for predicting the clasification based on the input
}
\description{
Building model for Spam Analysis.
}
\details{
This function creates a model depending on the input, and save it in the file
model.rda
}
\author{
Juan Manuel Hernandez
}

